Cole/get roster

+++ nhl-player-trigger/PlayerValue/Tables/PlayerValue.sql
@@ -2,5 +2,5 @@
 (
 	[id] INT NOT NULL PRIMARY KEY,
 	[name] VARCHAR(MAX) NOT NULL,
-	[value] DECIMAL(6, 3) NOT NULL
+	[value] FLOAT NOT NULL
 )
